cfs: a bit prettify cfs_shmem_size

+++ src/include/storage/cfs.h
@@ -133,7 +133,7 @@ int      cfs_msync(FileMap* map);
 FileMap* cfs_mmap(int md);
 int      cfs_munmap(FileMap* map);
 void     cfs_initialize(void);
-int      cfs_shmem_size(void);
+size_t   cfs_shmem_size(void);
 
 void     cfs_encrypt(const char* fname, void* block, uint32 offs, uint32 size);
 void     cfs_decrypt(const char* fname, void* block, uint32 offs, uint32 size);
